refactor: Describe fields with type hints in all doctypes (WIP)

+++ comfort/comfort_core/doctype/commission_range/commission_range.py
@@ -3,5 +3,5 @@ from frappe.model.document import Document
 
 class CommissionRange(Document):
     percentage: int
-    from_amount: float
-    to_amount: float
+    from_amount: int
+    to_amount: int
